+++ encodings/alp/src/alp_rd/compute/filter.rs
@@ -33,9 +33,10 @@ register_kernel!(FilterKernelAdapter(ALPRDVTable).lift());
 #[cfg(test)]
 mod test {
     use rstest::rstest;
-    use vortex_array::ToCanonical;
+    use vortex_array::{IntoArray, ToCanonical};
     use vortex_array::arrays::PrimitiveArray;
     use vortex_array::compute::filter;
+    use vortex_array::compute::conformance::filter::test_filter as test_filter_conformance;
     use vortex_array::validity::Validity;
     use vortex_buffer::buffer;
     use vortex_mask::Mask;
@@ -58,5 +59,33 @@ mod test {
             .to_primitive()
             .unwrap();
         assert_eq!(filtered.as_slice::<T>(), &[a, outlier]);
+    }
+
+    #[rstest]
+    #[case(0.1f32, 0.2f32, 3e25f32)]
+    #[case(0.1f64, 0.2f64, 3e100f64)]
+    fn test_filter_simple<T: ALPRDFloat>(#[case] a: T, #[case] b: T, #[case] outlier: T) {
+        test_filter_conformance(
+            &RDEncoder::new(&[a, b])
+                .encode(&PrimitiveArray::from_iter([a, b, outlier, b, outlier]))
+                .into_array(),
+        );
+    }
+
+    #[rstest]
+    #[case(0.1f32, 3e25f32)]
+    #[case(0.5f64, 1e100f64)]
+    fn test_filter_with_nulls<T: ALPRDFloat>(#[case] a: T, #[case] outlier: T) {
+        test_filter_conformance(
+            &RDEncoder::new(&[a])
+                .encode(&PrimitiveArray::from_option_iter([
+                    Some(a),
+                    None,
+                    Some(outlier),
+                    Some(a),
+                    None,
+                ]))
+                .into_array(),
+        );
     }
 }
