+++ encodings/alp/src/alp_rd/compute/mask.rs
@@ -29,7 +29,6 @@ mod tests {
     use rstest::rstest;
     use vortex_array::IntoArray;
     use vortex_array::arrays::PrimitiveArray;
-    use vortex_array::compute::conformance::filter::test_filter;
     use vortex_array::compute::conformance::mask::test_mask;
 
     use crate::{ALPRDFloat, RDEncoder};
@@ -50,34 +49,6 @@ mod tests {
     #[case(0.5f64, 1e100f64)]
     fn test_mask_with_nulls<T: ALPRDFloat>(#[case] a: T, #[case] outlier: T) {
         test_mask(
-            &RDEncoder::new(&[a])
-                .encode(&PrimitiveArray::from_option_iter([
-                    Some(a),
-                    None,
-                    Some(outlier),
-                    Some(a),
-                    None,
-                ]))
-                .into_array(),
-        );
-    }
-
-    #[rstest]
-    #[case(0.1f32, 0.2f32, 3e25f32)]
-    #[case(0.1f64, 0.2f64, 3e100f64)]
-    fn test_filter_simple<T: ALPRDFloat>(#[case] a: T, #[case] b: T, #[case] outlier: T) {
-        test_filter(
-            &RDEncoder::new(&[a, b])
-                .encode(&PrimitiveArray::from_iter([a, b, outlier, b, outlier]))
-                .into_array(),
-        );
-    }
-
-    #[rstest]
-    #[case(0.1f32, 3e25f32)]
-    #[case(0.5f64, 1e100f64)]
-    fn test_filter_with_nulls<T: ALPRDFloat>(#[case] a: T, #[case] outlier: T) {
-        test_filter(
             &RDEncoder::new(&[a])
                 .encode(&PrimitiveArray::from_option_iter([
                     Some(a),
