+++ vortex-array/src/arrays/varbin/compute/filter.rs
@@ -194,6 +194,7 @@ mod test {
     use crate::arrays::varbin::compute::filter::{
         filter_select_var_bin_by_index, filter_select_var_bin_by_slice,
     };
+    use crate::compute::conformance::filter::test_filter;
     use crate::validity::Validity;
 
     fn nullable_scalar_str(s: &str) -> Scalar {
@@ -309,5 +310,20 @@ mod test {
 
         assert_eq!(buf.scalar_at(0).unwrap(), null);
         assert_eq!(buf.scalar_at(1).unwrap(), null);
+    }
+
+    #[test]
+    fn test_filter_var_bin_array() {
+        let array = VarBinArray::from_vec(
+            vec!["hello", "world", "filter", "good", "bye"],
+            DType::Utf8(NonNullable),
+        );
+        test_filter(array.as_ref());
+
+        let array = VarBinArray::from_iter(
+            vec![Some("hello"), None, Some("filter"), Some("good"), None],
+            DType::Utf8(Nullable),
+        );
+        test_filter(array.as_ref());
     }
 }
