+++ vortex-array/src/arrays/varbinview/compute/filter.rs
@@ -15,3 +15,27 @@ impl FilterKernel for VarBinViewVTable {
 }
 
 register_kernel!(FilterKernelAdapter(VarBinViewVTable).lift());
+
+#[cfg(test)]
+mod tests {
+    use crate::arrays::VarBinViewArray;
+    use crate::compute::conformance::filter::test_filter;
+
+    #[test]
+    fn test_filter_var_bin_view_array() {
+        test_filter(
+            VarBinViewArray::from_iter_str(["one", "two", "three", "four", "five"]).as_ref(),
+        );
+
+        test_filter(
+            VarBinViewArray::from_iter_nullable_str([
+                Some("one"),
+                None,
+                Some("three"),
+                Some("four"),
+                Some("five"),
+            ])
+            .as_ref(),
+        );
+    }
+}
