+++ vortex-array/src/arrays/varbinview/compute/mod.rs
@@ -18,7 +18,6 @@ mod tests {
     use crate::array::Array;
     use crate::arrays::VarBinViewArray;
     use crate::canonical::ToCanonical;
-    use crate::compute::conformance::filter::test_filter;
     use crate::compute::conformance::mask::test_mask;
     use crate::compute::take;
 
@@ -53,24 +52,6 @@ mod tests {
         test_mask(VarBinViewArray::from_iter_str(["one", "two", "three", "four", "five"]).as_ref());
 
         test_mask(
-            VarBinViewArray::from_iter_nullable_str([
-                Some("one"),
-                None,
-                Some("three"),
-                Some("four"),
-                Some("five"),
-            ])
-            .as_ref(),
-        );
-    }
-
-    #[test]
-    fn test_filter_var_bin_view_array() {
-        test_filter(
-            VarBinViewArray::from_iter_str(["one", "two", "three", "four", "five"]).as_ref(),
-        );
-
-        test_filter(
             VarBinViewArray::from_iter_nullable_str([
                 Some("one"),
                 None,
